[Refactor] useMarkdownWidget composable
